refactor(core): split params + tool impls into directory modules (<1000)
Decompose four oversized tool files into directory modules:

- tools/params.rs (1910): split the *_params() schema fns into params/mod.rs
  + params/ext.rs.
- tools/gateway_tools.rs (1452): async impls → gateway_tools/async_impl.rs;
  sync entry points + helpers stay in mod.rs.
- tools/exo_ai.rs (1126): async impls/helpers → exo_ai/async_impl.rs; shared
  helpers + sync impl stay in mod.rs.
- tools/browser.rs (1092): the cfg(browser) CDP `mod real` → browser/real.rs.

All re-export via `pub use` so existing `params::…` / `gateway_tools::…` paths
are unchanged. Every resulting file is under 1000 lines. No behavior change.
